ethiccompanion-mvp/backend/requirements.txt
# Backend Requirements - Ethicompanion MVP

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP clients and API integration
httpx==0.25.2
aiohttp==3.9.1

# Google AI (Gemini) integration
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1

# Anthropic Claude integration
anthropic==0.7.8

# NeMo Guardrails (NVIDIA)
nemoguardrails==0.7.1

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Async and utilities
asyncio==3.4.3
typing-extensions==4.8.0

# Text processing and NLP
nltk==3.8.1
spacy==3.7.2
langchain==0.0.350
langchain-community==0.0.5

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Database (if needed for user data)
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# CORS and middleware
python-cors==1.7.0
